Project
Here, I want to describe that I have made the changes in coding because unfortunately I started the code for desktop. But now I modified the code and I applied all the coding for mobile. I also try to put the description of each block in content. Besides this, I also add the transition on "Read more" buttons. Moreover, I also use the Iframe for multimedia. Now I'm just writing the code for table in homepage and I just trying to apply CSS on that. After that I will use the media query for desktop. As well as my font family is also pending yet to apply.